requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
scipy==1.11.4
python-dotenv==1.0.0
requests>=2.32.2
pytest==7.4.3
matplotlib>=3.8.0
seaborn>=0.13.0
google-generativeai>=0.3.0
openpyxl>=3.1.0
python-multipart>=0.0.6 